pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "churn-prediction-platform"
version = "0.1.0"
description = "Production-ready customer churn prediction platform with MLOps capabilities"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    # Core Data Science
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    
    # Feature Engineering & Time Series
    "pyarrow>=12.0.0",
    "fastparquet>=2023.4.0",
    
    # Explainability
    "shap>=0.42.0",
    
    # Experiment Tracking & ML Ops
    "mlflow>=2.8.0",
    
    # API & Web
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Configuration
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    
    # Monitoring & Drift Detection
    "scipy>=1.11.0",
    
    # Visualization & Reporting
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.17.0",
    
    # CLI
    "click>=8.1.0",
    "rich>=13.0.0",
    
    # Database & Storage
    "sqlalchemy>=2.0.0",
    
    # Utilities
    "python-dateutil>=2.8.0",
    "tqdm>=4.65.0",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[project.scripts]
churn-train = "src.training_pipeline.run:main"
churn-infer = "src.inference.batch:main"
churn-monitor = "src.monitoring.drift_detector:main"
churn-api = "src.inference.api:start_server"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing"
